+++ src/main/scala/dto/entities/TopicDto.scala
@@ -3,4 +3,4 @@ package dto.entities
 import java.sql.Timestamp
 
 
-case class TopicDto(subject: String, lastPostTimestamp: Timestamp, topicId: Option[Int] = None)
+case class TopicDto(subject: String, lastPostTimestamp: Timestamp, topicId: Option[Long] = None)
